Changed the way optionalFrom<T> works. What it previously did was: a) Try to find the JSON value for the given key/keypath, if it could not find it, it'd pass nil to the tranformation which could then do whatever it liked, including either returning nil, returning something or throwing. If the transformation threw it'd get ignored and the whole thing would return nil. b) Try to find the JSON value for the given key/keypath, if it could find it, it'd pass it to the transformation, which could then do whatever it liked, including either returning nil, returning something or throwing. If the transform threw it'd get ignored and the whole thing would return nil.
It was very confusing and didn't really have a clear purpose.

What it does now:
I've made it work like the `map` function in that, when we find a value in the JSON we pass that to the transformation and the transformation can transform that value into any non-optional value than we return that or if it throws we rethrow that error. This function now only returns nil when the JSON value does not exist. I will be adding a NilConvertible protocol in a later commit which will be used in the specific case where the consumer wants to convert a nil value into a non-optional value.